more generics...
git-svn-id: file:///home/coren/project/backup_ho/trunk/Core@524 b8de521a-fa13-0410-8113-fc7ffabbd59a

+++ src/java/de/hattrickorganizer/credits/Ball.java
@@ -74,15 +74,15 @@ public class Ball {
         }
 
         if (goRight) {
-            x += ((float) vergangeneZeit / 5.0f);
+            x += (vergangeneZeit / 5.0f);
         } else {
-            x -= ((float) vergangeneZeit / 5.0f);
+            x -= (vergangeneZeit / 5.0f);
         }
 
         if (goDown) {
-            y += ((float) vergangeneZeit / 5.0f);
+            y += (vergangeneZeit / 5.0f);
         } else {
-            y -= ((float) vergangeneZeit / 5.0f);
+            y -= (vergangeneZeit / 5.0f);
         }
     }
